+++ gear/file_distribution/src/python/file_distribution_app/run.py
@@ -7,7 +7,7 @@ from flywheel.rest import ApiException
 from fw_gear import GearContext
 from gear_execution.gear_execution import (
     ClientWrapper,
-    ContextClient,
+    GearBotClient,
     GearEngine,
     GearExecutionEnvironment,
     GearExecutionError,
@@ -32,8 +32,9 @@ class FileDistributionVisitor(GearExecutionEnvironment):
         self,
         client: ClientWrapper,
         file_input: InputFileWrapper,
-        target_project: str,
-        batch_size: int,
+        target_project: Optional[str] = None,
+        staging_project_id: Optional[str] = None,
+        batch_size: int = 8,
         downstream_gears: Optional[List[str]] = None,
         include: Optional[str] = None,
         exclude: Optional[str] = None,
@@ -42,6 +43,7 @@ class FileDistributionVisitor(GearExecutionEnvironment):
 
         self.__file_input = file_input
         self.__target_project = target_project
+        self.__staging_project_id = staging_project_id
         self.__include = include
         self.__exclude = exclude
         self.__batch_size = batch_size
@@ -68,18 +70,22 @@ class FileDistributionVisitor(GearExecutionEnvironment):
           GearExecutionError if any expected inputs are missing
         """
 
-        client = ContextClient.create(context=context)
+        client = GearBotClient.create(context=context, parameter_store=parameter_store)
         file_input = InputFileWrapper.create(input_name="input_file", context=context)
         if not file_input:
             raise GearExecutionError("No input file provided")
 
         options = context.config.opts
         target_project = options.get("target_project", None)
-        if not target_project:
-            raise GearExecutionError("No target project provided")
+        staging_project_id = options.get("staging_project_id", None)
+
+        if not target_project and not staging_project_id:
+            raise GearExecutionError(
+                "One of target_project or staging_project_id must be provided"
+            )
 
         # for scheduling
-        batch_size = options.get("batch_size", 1)
+        batch_size = options.get("batch_size", 8)
         downstream_gears = parse_string_to_list(options.get("downstream_gears", ""))
 
         try:
@@ -96,6 +102,7 @@ class FileDistributionVisitor(GearExecutionEnvironment):
             client=client,
             file_input=file_input,
             target_project=target_project,
+            staging_project_id=staging_project_id,
             batch_size=batch_size,
             downstream_gears=downstream_gears,
             include=options.get("include", None),
@@ -117,9 +124,10 @@ class FileDistributionVisitor(GearExecutionEnvironment):
             proxy=self.proxy,
             error_writer=ListErrorWriter(container_id=file_id, fw_path=fw_path),
             file=file,
-            target_project=self.__target_project,
             centers=self.__centers,
             batch_size=self.__batch_size,
+            target_project=self.__target_project,
+            staging_project_id=self.__staging_project_id,
             downstream_gears=self.__downstream_gears,
         )
 
@@ -127,7 +135,7 @@ class FileDistributionVisitor(GearExecutionEnvironment):
 def main():
     """Main method for File Distribution."""
 
-    GearEngine().run(gear_type=FileDistributionVisitor)
+    GearEngine.create_with_parameter_store().run(gear_type=FileDistributionVisitor)
 
 
 if __name__ == "__main__":