+++ gear/gather_submission_status/src/python/gather_submission_status_app/run.py
@@ -161,7 +161,7 @@ class GatherSubmissionStatusVisitor(GearExecutionEnvironment):
                 data=error_writer.errors().model_dump(by_alias=True),
             )
 
-            gear_name = self.gear_name(context, "gather-submission-status")
+            gear_name = self.get_gear_name(context, "gather-submission-status")
             context.metadata.add_file_tags(self.__file_input.file_input, tags=gear_name)