+++ gear/identifier_lookup/src/python/identifier_app/run.py
@@ -8,7 +8,10 @@ from pathlib import Path
 from typing import List, Literal, Optional, TextIO
 
 from botocore.exceptions import ClientError
-from configs.ingest_configs import ModuleConfigs
+from configs.ingest_configs import (
+    ModuleConfigs,
+    load_form_ingest_configurations,
+)
 from error_logging.error_logger import ErrorLogTemplate
 from error_logging.qc_status_log_creator import (
     FileVisitAnnotator,
@@ -43,7 +46,6 @@ from nacc_common.field_names import FieldNames
 from outputs.error_writer import ListErrorWriter
 from pydantic import ValidationError
 from s3.s3_bucket import S3BucketInterface
-from utils.utils import load_form_ingest_configurations
 
 log = logging.getLogger(__name__)
 
@@ -108,7 +110,7 @@ class IdentifierLookupVisitor(GearExecutionEnvironment):
         preserve_case = options.get("preserve_case", False)
         module = options.get("module")
         single_center = options.get("single_center", True)
-        gear_name = GearExecutionEnvironment.gear_name(context, "identifer-lookup")
+        gear_name = GearExecutionEnvironment.get_gear_name(context, "identifer-lookup")
 
         # Note: form_configs_file is optional for 'nacc' direction
         # When not provided, only basic identifier lookup will be performed